+++ src/mlw/eval_match.ml
@@ -173,6 +173,8 @@ let rec eval_match kn stop env t =
   t_attr_copy t (match t.t_node with
     | Tapp (ls, [t1;t2]) when ls_equal ls ps_equ ->
         cs_equ env (eval env t1) (eval env t2)
+    | Tapp (ls, [_]) when ls_equal ls Theory.well_formed ->
+        Term.t_true
     | Tnot { t_node = Tapp (ls, [t1;t2]) } when ls_equal ls ps_equ ->
         t_not_simp (cs_equ env (eval env t1) (eval env t2))
     | Tapp (ls, [t1]) when is_projection ls ->